requirements.txt
requests>=2.31.0
playwright>=1.42.0
eventlet>=0.35.0
schedule>=1.2.0
cf_clearance>=0.3.0
biliup>=0.1.0
xhs>=0.2.13
qrcode>=7.4.0
loguru>=0.7.2
flask>=3.0.0
gevent>=24.2.1
pydantic>=2.6.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
watchdog>=3.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
gradio>=5.15.0
aiosqlite>=0.19.0